reorganize custom types
